Move `grpc` to `plugin`
This allows vanity to depend on `plugin/grpc` in go1.5beta1.
